requirements.txt
transformers
torch
torchaudio
soundfile
numpy
Pillow
gtts
gradio
